refactor: 픽스쳐 디렉토리 변경
