+++ src/test/java/kitchenpos/fixture/MenuFixture.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.fixture;
+package kitchenpos.fixture;
 
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuProduct;
