+++ src/test/java/kitchenpos/fixture/OrderLineItemFixture.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.fixture;
+package kitchenpos.fixture;
 
 import kitchenpos.domain.OrderLineItem;
