+++ src/test/java/kitchenpos/ui/MenuRestControllerTest.java
@@ -22,7 +22,7 @@ import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
-import static kitchenpos.application.fixture.MenuFixture.*;
+import static kitchenpos.fixture.MenuFixture.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
